Plugin with gameplay tag native with namespace association to ArcPlugGameplayTags
