+++ Unreal/ArcPlus/Source/ArcPlus/Private/ArcPlusGameplayTags.cpp
@@ -1,0 +1,12 @@
+﻿// Copyright Broken Rock Studios LLC. All Rights Reserved.
+
+
+#include "ArcPlusGameplayTags.h"
+#include "NativeGameplayTags.h"
+
+
+
+namespace ArcPlusGameplayTags
+{
+	UE_DEFINE_GAMEPLAY_TAG_COMMENT(Inventory_Action_SwappableCheck, "Inventory.Action.SwappableCheck", "Used for checking Inventory Swap Checks")
+}
