+++ Unreal/ArcPlus/Source/ArcPlus/Public/ArcPlusGameplayTags.h
@@ -1,0 +1,12 @@
+﻿// Copyright Broken Rock Studios LLC. All Rights Reserved.
+
+#pragma once
+
+#include "CoreMinimal.h"
+#include "NativeGameplayTags.h"
+
+namespace ArcPlusGameplayTags
+{
+	UE_DECLARE_GAMEPLAY_TAG_EXTERN(Inventory_Action_SwappableCheck)
+}
+
